+++ api/http/api.go
@@ -149,6 +149,7 @@ func (self *HttpServer) Serve(listener net.Listener) {
 	self.registerEndpoint("get", "/interfaces", self.listInterfaces)
 
 	// cluster config endpoints
+	self.registerEndpoint("get", "/cluster/configuration", self.getClusterConfiguration)
 	self.registerEndpoint("get", "/cluster/servers", self.listServers)
 	self.registerEndpoint("del", "/cluster/servers/:id", self.removeServers)
 	self.registerEndpoint("post", "/cluster/shards", self.createShard)
@@ -157,6 +158,7 @@ func (self *HttpServer) Serve(listener net.Listener) {
 	self.registerEndpoint("get", "/cluster/shard_spaces", self.getShardSpaces)
 	self.registerEndpoint("post", "/cluster/shard_spaces/:db", self.createShardSpace)
 	self.registerEndpoint("del", "/cluster/shard_spaces/:db/:name", self.dropShardSpace)
+	self.registerEndpoint("post", "/cluster/shard_spaces/:db/:name", self.updateShardSpace)
 	self.registerEndpoint("post", "/cluster/database_configs/:db", self.configureDatabase)
 
 	// return whether the cluster is in sync or not
@@ -431,6 +433,11 @@ func (self *HttpServer) createDatabase(w libhttp.ResponseWriter, r *libhttp.Requ
 		if err != nil {
 			return libhttp.StatusBadRequest, err.Error()
 		}
+		if !isValidDbName(createRequest.Name) {
+			m := "Unable to create database without name"
+			log.Error(m)
+			return libhttp.StatusBadRequest, m
+		}
 		err = self.coordinator.CreateDatabase(user, createRequest.Name)
 		if err != nil {
 			log.Error("Cannot create database %s. Error: %s", createRequest.Name, err)
@@ -439,6 +446,10 @@ func (self *HttpServer) createDatabase(w libhttp.ResponseWriter, r *libhttp.Requ
 		log.Debug("Created database %s", createRequest.Name)
 		return libhttp.StatusCreated, nil
 	})
+}
+
+func isValidDbName(name string) bool {
+	return strings.TrimSpace(name) != ""
 }
 
 func (self *HttpServer) dropDatabase(w libhttp.ResponseWriter, r *libhttp.Request) {
@@ -1183,5 +1194,35 @@ func (self *HttpServer) configureDatabase(w libhttp.ResponseWriter, r *libhttp.R
 			}
 		}
 		return libhttp.StatusCreated, nil
+	})
+}
+
+func (self *HttpServer) updateShardSpace(w libhttp.ResponseWriter, r *libhttp.Request) {
+	self.tryAsClusterAdmin(w, r, func(u User) (int, interface{}) {
+		space := &cluster.ShardSpace{}
+		decoder := json.NewDecoder(r.Body)
+		err := decoder.Decode(space)
+		if err != nil {
+			return libhttp.StatusInternalServerError, err.Error()
+		}
+		space.Database = r.URL.Query().Get(":db")
+		space.Name = r.URL.Query().Get(":name")
+		if !self.clusterConfig.DatabaseExists(space.Database) {
+			return libhttp.StatusNotAcceptable, "Can't update a shard space for a database that doesn't exist"
+		}
+		if !self.clusterConfig.ShardSpaceExists(space) {
+			return libhttp.StatusNotAcceptable, "Can't update a shard space that doesn't exist"
+		}
+
+		if err := self.raftServer.UpdateShardSpace(space); err != nil {
+			return libhttp.StatusInternalServerError, err.Error()
+		}
+		return libhttp.StatusOK, nil
+	})
+}
+
+func (self *HttpServer) getClusterConfiguration(w libhttp.ResponseWriter, r *libhttp.Request) {
+	self.tryAsClusterAdmin(w, r, func(u User) (int, interface{}) {
+		return libhttp.StatusOK, self.clusterConfig.SerializableConfiguration()
 	})
 }