+++ cluster/cluster_configuration.go
@@ -125,16 +125,16 @@ func NewClusterConfiguration(
 	}
 }
 
-func (self *ClusterConfiguration) DoesShardSpaceExist(space *ShardSpace) error {
+func (self *ClusterConfiguration) ShardSpaceExists(space *ShardSpace) bool {
 	self.shardLock.RLock()
 	defer self.shardLock.RUnlock()
 	dbSpaces := self.databaseShardSpaces[space.Database]
 	for _, s := range dbSpaces {
 		if s.Name == space.Name {
-			return fmt.Errorf("Shard space %s exists", space.Name)
+			return true
 		}
 	}
-	return nil
+	return false
 }
 
 func (self *ClusterConfiguration) SetShardCreator(shardCreator ShardCreator) {
@@ -159,6 +159,12 @@ func (self *ClusterConfiguration) GetShardSpaces() []*ShardSpace {
 		spaces = append(spaces, databaseSpaces...)
 	}
 	return spaces
+}
+
+func (self *ClusterConfiguration) GetShardSpacesForDatabase(database string) []*ShardSpace {
+	self.shardLock.RLock()
+	defer self.shardLock.RUnlock()
+	return self.databaseShardSpaces[database]
 }
 
 // called by the server, this will wake up every 10 mintues to see if it should
@@ -611,6 +617,17 @@ type SavedConfiguration struct {
 
 func (self *ClusterConfiguration) Save() ([]byte, error) {
 	log.Debug("Dumping the cluster configuration")
+	data := self.SerializableConfiguration()
+	b := bytes.NewBuffer(nil)
+	err := gob.NewEncoder(b).Encode(&data)
+	if err != nil {
+		return nil, err
+	}
+
+	return b.Bytes(), nil
+}
+
+func (self *ClusterConfiguration) SerializableConfiguration() *SavedConfiguration {
 	data := &SavedConfiguration{
 		Databases:           make(map[string]uint8, len(self.DatabaseReplicationFactors)),
 		Admins:              self.clusterAdmins,
@@ -626,14 +643,7 @@ func (self *ClusterConfiguration) Save() ([]byte, error) {
 	for k := range self.DatabaseReplicationFactors {
 		data.Databases[k] = 0
 	}
-
-	b := bytes.NewBuffer(nil)
-	err := gob.NewEncoder(b).Encode(&data)
-	if err != nil {
-		return nil, err
-	}
-
-	return b.Bytes(), nil
+	return data
 }
 
 func (self *ClusterConfiguration) convertShardsToNewShardData(shards []*ShardData) []*NewShardData {
@@ -787,18 +797,6 @@ func (self *ClusterConfiguration) LastContinuousQueryRunTime() time.Time {
 
 func (self *ClusterConfiguration) SetLastContinuousQueryRunTime(t time.Time) {
 	self.continuousQueryTimestamp = t
-}
-
-func (self *ClusterConfiguration) GetMapForJsonSerialization() map[string]interface{} {
-	jsonObject := make(map[string]interface{})
-	dbs := make([]string, 0)
-	for db := range self.DatabaseReplicationFactors {
-		dbs = append(dbs, db)
-	}
-	jsonObject["databases"] = dbs
-	jsonObject["cluster_admins"] = self.clusterAdmins
-	jsonObject["database_users"] = self.dbUsers
-	return jsonObject
 }
 
 func (self *ClusterConfiguration) createDefaultShardSpace(database string) (*ShardSpace, error) {
@@ -1311,6 +1309,17 @@ func (self *ClusterConfiguration) AddShardSpace(space *ShardSpace) error {
 	copy(newSpaces[1:], databaseSpaces)
 	newSpaces[0] = space
 	self.databaseShardSpaces[space.Database] = newSpaces
+	return nil
+}
+
+func (self *ClusterConfiguration) UpdateShardSpace(space *ShardSpace) error {
+	self.shardLock.Lock()
+	defer self.shardLock.Unlock()
+	oldSpace := self.getShardSpaceByDatabaseAndName(space.Database, space.Name)
+	if oldSpace == nil {
+		return nil
+	}
+	oldSpace.UpdateFromSpace(space)
 	return nil
 }