+++ datastore/shard.go
@@ -1,8 +1,6 @@
 package datastore
 
 import (
-	"bytes"
-	"encoding/binary"
 	"errors"
 	"fmt"
 	"math"
@@ -61,20 +59,12 @@ func (self *Shard) Write(database string, series []*protocol.Series) error {
 				// keyBuffer and dataBuffer have to be recreated since we are
 				// batching the writes, otherwise new writes will override the
 				// old writes that are still in memory
-				keyBuffer := bytes.NewBuffer(make([]byte, 0, 24))
 				dataBuffer := proto.NewBuffer(nil)
 				var err error
 
-				binary.Write(keyBuffer, binary.BigEndian, &id)
-				timestamp := self.convertTimestampToUint(point.GetTimestampInMicroseconds())
-				// pass the uint64 by reference so binary.Write() doesn't create a new buffer
-				// see the source code for intDataSize() in binary.go
-				binary.Write(keyBuffer, binary.BigEndian, &timestamp)
-				binary.Write(keyBuffer, binary.BigEndian, point.SequenceNumber)
-				pointKey := keyBuffer.Bytes()
-
+				sk := newStorageKey(id, point.GetTimestamp(), point.GetSequenceNumber())
 				if point.Values[fieldIndex].GetIsNull() {
-					wb = append(wb, storage.Write{Key: pointKey, Value: nil})
+					wb = append(wb, storage.Write{Key: sk.bytes(), Value: nil})
 					goto check
 				}
 
@@ -82,7 +72,7 @@ func (self *Shard) Write(database string, series []*protocol.Series) error {
 				if err != nil {
 					return err
 				}
-				wb = append(wb, storage.Write{Key: pointKey, Value: dataBuffer.Bytes()})
+				wb = append(wb, storage.Write{Key: sk.bytes(), Value: dataBuffer.Bytes()})
 			check:
 				count++
 				if count >= self.writeBatchSize {
@@ -145,179 +135,57 @@ func (self *Shard) IsClosed() bool {
 }
 
 func (self *Shard) executeQueryForSeries(querySpec *parser.QuerySpec, seriesName string, columns []string, processor engine.Processor) error {
-	startTimeBytes := self.byteArrayForTime(querySpec.GetStartTime())
-	endTimeBytes := self.byteArrayForTime(querySpec.GetEndTime())
-
 	fields, err := self.getFieldsForSeries(querySpec.Database(), seriesName, columns)
 	if err != nil {
 		log.Error("Error looking up fields for %s: %s", seriesName, err)
 		return err
 	}
 
-	fieldCount := len(fields)
-	rawColumnValues := make([]rawColumnValue, fieldCount, fieldCount)
+	if querySpec.IsSinglePointQuery() {
+		log.Debug("Running single query for series %s, fields %v", seriesName, fields)
+		return self.executeSinglePointQuery(querySpec, seriesName, fields, processor)
+	}
+
+	startTime := querySpec.GetStartTime()
+	endTime := querySpec.GetEndTime()
+
 	query := querySpec.SelectQuery()
 
 	aliases := query.GetTableAliases(seriesName)
-	if querySpec.IsSinglePointQuery() {
-		series, err := self.fetchSinglePoint(querySpec, seriesName, fields)
-		if err != nil {
-			log.Error("Error reading a single point: %s", err)
-			return err
-		}
-		if len(series.Points) > 0 {
-			_, err := processor.Yield(series)
-			return err
-		}
-		return nil
-	}
 
-	fieldNames, iterators := self.getIterators(fields, startTimeBytes, endTimeBytes, query.Ascending)
-	defer func() {
-		for _, it := range iterators {
-			it.Close()
-		}
-	}()
-
+	fieldNames, iterators := self.getIterators(fields, startTime, endTime, query.Ascending)
 	seriesOutgoing := &protocol.Series{Name: protocol.String(seriesName), Fields: fieldNames, Points: make([]*protocol.Point, 0, self.pointBatchSize)}
+	pi := NewPointIterator(iterators, fields, querySpec.GetStartTime(), querySpec.GetEndTime(), query.Ascending)
+	defer pi.Close()
 
-	// TODO: clean up, this is super gnarly
-	// optimize for the case where we're pulling back only a single column or aggregate
-	buffer := bytes.NewBuffer(nil)
-	valueBuffer := proto.NewBuffer(nil)
-	for {
-		isValid := false
-		point := &protocol.Point{Values: make([]*protocol.FieldValue, fieldCount, fieldCount)}
-
-		for i, it := range iterators {
-			if rawColumnValues[i].value != nil || !it.Valid() {
-				if err := it.Error(); err != nil {
-					return err
-				}
-				continue
-			}
-
-			key := it.Key()
-			if len(key) < 16 {
-				continue
-			}
-
-			if !isPointInRange(fields[i].IdAsBytes(), startTimeBytes, endTimeBytes, key) {
-				continue
-			}
-
-			value := it.Value()
-			sequenceNumber := key[16:]
-
-			rawTime := key[8:16]
-			rawColumnValues[i] = rawColumnValue{time: rawTime, sequence: sequenceNumber, value: value}
-		}
-
-		var pointTimeRaw []byte
-		var pointSequenceRaw []byte
-		// choose the highest (or lowest in case of ascending queries) timestamp
-		// and sequence number. that will become the timestamp and sequence of
-		// the next point.
-		for _, value := range rawColumnValues {
-			if value.value == nil {
-				continue
-			}
-
-			pointTimeRaw, pointSequenceRaw = value.updatePointTimeAndSequence(pointTimeRaw,
-				pointSequenceRaw, query.Ascending)
-		}
-
-		for i, iterator := range iterators {
-			// if the value is nil or doesn't match the point's timestamp and sequence number
-			// then skip it
-			if rawColumnValues[i].value == nil ||
-				!bytes.Equal(rawColumnValues[i].time, pointTimeRaw) ||
-				!bytes.Equal(rawColumnValues[i].sequence, pointSequenceRaw) {
-
-				point.Values[i] = &protocol.FieldValue{IsNull: &TRUE}
-				continue
-			}
-
-			// if we emitted at lease one column, then we should keep
-			// trying to get more points
-			isValid = true
-
-			// advance the iterator to read a new value in the next iteration
-			if query.Ascending {
-				iterator.Next()
-			} else {
-				iterator.Prev()
-			}
-
-			fv := &protocol.FieldValue{}
-			valueBuffer.SetBuf(rawColumnValues[i].value)
-			err := valueBuffer.Unmarshal(fv)
-			if err != nil {
-				log.Error("Error while running query: %s", err)
-				return err
-			}
-			point.Values[i] = fv
-			rawColumnValues[i].value = nil
-		}
-
-		var sequence uint64
-		var t uint64
-
-		// set the point sequence number and timestamp
-		buffer.Reset()
-		buffer.Write(pointSequenceRaw)
-		binary.Read(buffer, binary.BigEndian, &sequence)
-		buffer.Reset()
-		buffer.Write(pointTimeRaw)
-		binary.Read(buffer, binary.BigEndian, &t)
-
-		time := self.convertUintTimestampToInt64(&t)
-		point.SetTimestampInMicroseconds(time)
-		point.SequenceNumber = &sequence
-
-		// stop the loop if we ran out of points
-		if !isValid {
-			break
-		}
-
-		shouldContinue := true
-
-		seriesOutgoing.Points = append(seriesOutgoing.Points, point)
-
+	for pi.Valid() {
+		p := pi.Point()
+		seriesOutgoing.Points = append(seriesOutgoing.Points, p)
 		if len(seriesOutgoing.Points) >= self.pointBatchSize {
-			for _, alias := range aliases {
-				series := &protocol.Series{
-					Name:   proto.String(alias),
-					Fields: fieldNames,
-					Points: seriesOutgoing.Points,
-				}
-				if ok, err := processor.Yield(series); !ok || err != nil {
-					log.Debug("Stopping processing.")
-					shouldContinue = false
-					if err != nil {
-						log.Error("Error while processing data: %v", err)
-						return err
-					}
+			ok, err := yieldToProcessor(seriesOutgoing, processor, aliases)
+			if !ok || err != nil {
+				log.Debug("Stopping processing.")
+				if err != nil {
+					log.Error("Error while processing data: %v", err)
+					return err
 				}
 			}
 			seriesOutgoing = &protocol.Series{Name: protocol.String(seriesName), Fields: fieldNames, Points: make([]*protocol.Point, 0, self.pointBatchSize)}
 		}
 
-		if !shouldContinue {
-			break
-		}
+		pi.Next()
+	}
+
+	if err := pi.Error(); err != nil {
+		return err
 	}
 
 	//Yield remaining data
-	for _, alias := range aliases {
-		log.Debug("Final Flush %s", alias)
-		series := &protocol.Series{Name: protocol.String(alias), Fields: seriesOutgoing.Fields, Points: seriesOutgoing.Points}
-		if ok, err := processor.Yield(series); !ok || err != nil {
-			log.Debug("Cancelled...")
-			if err != nil {
-				log.Error("Error while processing data: %v", err)
-				return err
-			}
+	if ok, err := yieldToProcessor(seriesOutgoing, processor, aliases); !ok || err != nil {
+		log.Debug("Stopping processing remaining points...")
+		if err != nil {
+			log.Error("Error while processing data: %v", err)
+			return err
 		}
 	}
 
@@ -355,42 +223,22 @@ func (self *Shard) DropFields(fields []*metastore.Field) error {
 	if self.closed {
 		return fmt.Errorf("Shard is closed")
 	}
-	startTimeBytes := []byte{0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00}
-	endTimeBytes := []byte{0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF}
-	return self.deleteRangeOfFields(fields, startTimeBytes, endTimeBytes)
+	return self.deleteRangeOfFields(fields, math.MinInt64, math.MaxInt64)
 }
 
-func (self *Shard) byteArrayForTimeInt(time int64) []byte {
-	timeBuffer := bytes.NewBuffer(make([]byte, 0, 8))
-	binary.Write(timeBuffer, binary.BigEndian, self.convertTimestampToUint(&time))
-	bytes := timeBuffer.Bytes()
-	return bytes
-}
-
-func (self *Shard) byteArraysForStartAndEndTimes(startTime, endTime int64) ([]byte, []byte) {
-	return self.byteArrayForTimeInt(startTime), self.byteArrayForTimeInt(endTime)
-}
-
-func (self *Shard) deleteRangeOfSeriesCommon(database, series string, startTimeBytes, endTimeBytes []byte) error {
+func (self *Shard) deleteRangeOfSeries(database, series string, startTime, endTime time.Time) error {
 	fields := self.metaStore.GetFieldsForSeries(database, series)
-	return self.deleteRangeOfFields(fields, startTimeBytes, endTimeBytes)
+	st := common.TimeToMicroseconds(startTime)
+	et := common.TimeToMicroseconds(endTime)
+	return self.deleteRangeOfFields(fields, st, et)
 }
 
-func (self *Shard) deleteRangeOfFields(fields []*metastore.Field, startTimeBytes, endTimeBytes []byte) error {
-	startKey := bytes.NewBuffer(nil)
-	endKey := bytes.NewBuffer(nil)
+func (self *Shard) deleteRangeOfFields(fields []*metastore.Field, st, et int64) error {
 	for _, field := range fields {
-		idBytes := field.IdAsBytes()
-		startKey.Reset()
-		startKey.Write(idBytes)
-		startKey.Write(startTimeBytes)
-		startKey.Write([]byte{0, 0, 0, 0, 0, 0, 0, 0})
-		endKey.Reset()
-		endKey.Write(idBytes)
-		endKey.Write(endTimeBytes)
-		endKey.Write([]byte{0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff})
+		sk := newStorageKey(field.Id, st, 0)
+		ek := newStorageKey(field.Id, et, maxSeqNumber)
 
-		err := self.db.Del(startKey.Bytes(), endKey.Bytes())
+		err := self.db.Del(sk.bytes(), ek.bytes())
 		if err != nil {
 			return err
 		}
@@ -403,11 +251,6 @@ func (self *Shard) deleteRangeOfFields(fields []*metastore.Field, startTimeBytes
 // 	self.db.CompactRange(levigo.Range{})
 // 	log.Info("Shard compaction is done")
 // }
-
-func (self *Shard) deleteRangeOfSeries(database, series string, startTime, endTime time.Time) error {
-	startTimeBytes, endTimeBytes := self.byteArraysForStartAndEndTimes(common.TimeToMicroseconds(startTime), common.TimeToMicroseconds(endTime))
-	return self.deleteRangeOfSeriesCommon(database, series, startTimeBytes, endTimeBytes)
-}
 
 func (self *Shard) deleteRangeOfRegex(database string, regex *regexp.Regexp, startTime, endTime time.Time) error {
 	series := self.metaStore.GetSeriesForDatabaseAndRegex(database, regex)
@@ -431,13 +274,6 @@ func (self *Shard) hasReadAccess(querySpec *parser.QuerySpec) bool {
 	return true
 }
 
-func (self *Shard) byteArrayForTime(t time.Time) []byte {
-	timeBuffer := bytes.NewBuffer(make([]byte, 0, 8))
-	timeMicro := common.TimeToMicroseconds(t)
-	binary.Write(timeBuffer, binary.BigEndian, self.convertTimestampToUint(&timeMicro))
-	return timeBuffer.Bytes()
-}
-
 func (self *Shard) close() {
 	self.closeLock.Lock()
 	defer self.closeLock.Unlock()
@@ -446,14 +282,7 @@ func (self *Shard) close() {
 	self.db = nil
 }
 
-func (self *Shard) convertTimestampToUint(t *int64) uint64 {
-	if *t < 0 {
-		return uint64(math.MaxInt64 + *t + 1)
-	}
-	return uint64(*t) + uint64(math.MaxInt64) + uint64(1)
-}
-
-func (self *Shard) fetchSinglePoint(querySpec *parser.QuerySpec, series string, fields []*metastore.Field) (*protocol.Series, error) {
+func (self *Shard) executeSinglePointQuery(querySpec *parser.QuerySpec, series string, fields []*metastore.Field, p engine.Processor) error {
 	query := querySpec.SelectQuery()
 	fieldCount := len(fields)
 	fieldNames := make([]string, 0, fieldCount)
@@ -461,59 +290,64 @@ func (self *Shard) fetchSinglePoint(querySpec *parser.QuerySpec, series string, 
 	timestamp := common.TimeToMicroseconds(query.GetStartTime())
 	sequenceNumber, err := query.GetSinglePointQuerySequenceNumber()
 	if err != nil {
-		return nil, err
+		return err
 	}
 
-	timeAndSequenceBuffer := bytes.NewBuffer(make([]byte, 0, 16))
-	binary.Write(timeAndSequenceBuffer, binary.BigEndian, self.convertTimestampToUint(&timestamp))
-	binary.Write(timeAndSequenceBuffer, binary.BigEndian, sequenceNumber)
-	sequenceNumber_uint64 := uint64(sequenceNumber)
-	point.SequenceNumber = &sequenceNumber_uint64
+	// set the timestamp and sequence number
+	point.SequenceNumber = &sequenceNumber
 	point.SetTimestampInMicroseconds(timestamp)
 
-	timeAndSequenceBytes := timeAndSequenceBuffer.Bytes()
 	for _, field := range fields {
-		pointKeyBuff := bytes.NewBuffer(make([]byte, 0, 24))
-		pointKeyBuff.Write(field.IdAsBytes())
-		pointKeyBuff.Write(timeAndSequenceBytes)
-
-		if data, err := self.db.Get(pointKeyBuff.Bytes()); err != nil {
-			return nil, err
-		} else {
-			fieldValue := &protocol.FieldValue{}
-			err := proto.Unmarshal(data, fieldValue)
-			if err != nil {
-				return nil, err
-			}
-			if data != nil {
-				fieldNames = append(fieldNames, field.Name)
-				point.Values = append(point.Values, fieldValue)
-			}
+		sk := newStorageKey(field.Id, timestamp, sequenceNumber)
+		data, err := self.db.Get(sk.bytes())
+		if err != nil {
+			return err
 		}
+
+		if data == nil {
+			continue
+		}
+
+		fieldValue := &protocol.FieldValue{}
+		err = proto.Unmarshal(data, fieldValue)
+		if err != nil {
+			return err
+		}
+		fieldNames = append(fieldNames, field.Name)
+		point.Values = append(point.Values, fieldValue)
 	}
 
 	result := &protocol.Series{Name: &series, Fields: fieldNames, Points: []*protocol.Point{point}}
 
-	return result, nil
+	if len(result.Points) > 0 {
+		_, err := p.Yield(result)
+		return err
+	}
+	return nil
 }
 
-func (self *Shard) getIterators(fields []*metastore.Field, start, end []byte, isAscendingQuery bool) (fieldNames []string, iterators []storage.Iterator) {
+func (self *Shard) getIterators(fields []*metastore.Field, start, end time.Time, isAscendingQuery bool) (fieldNames []string, iterators []storage.Iterator) {
 	iterators = make([]storage.Iterator, len(fields))
 	fieldNames = make([]string, len(fields))
 
 	// start the iterators to go through the series data
 	for i, field := range fields {
-		idBytes := field.IdAsBytes()
 		fieldNames[i] = field.Name
 		iterators[i] = self.db.Iterator()
 
-		if isAscendingQuery {
-			iterators[i].Seek(append(idBytes, start...))
-		} else {
-			iterators[i].Seek(append(append(idBytes, end...), MAX_SEQUENCE...))
-			if iterators[i].Valid() {
-				iterators[i].Prev()
-			}
+		t := start
+		var seq uint64 = 0
+		if !isAscendingQuery {
+			t = end
+			seq = maxSeqNumber
+		}
+
+		tmicro := common.TimeToMicroseconds(t)
+		sk := newStorageKey(field.Id, tmicro, seq)
+		iterators[i].Seek(sk.bytes())
+
+		if !isAscendingQuery && iterators[i].Valid() {
+			iterators[i].Prev()
 		}
 
 		if err := iterators[i].Error(); err != nil {
@@ -522,13 +356,6 @@ func (self *Shard) getIterators(fields []*metastore.Field, start, end []byte, is
 		}
 	}
 	return
-}
-
-func (self *Shard) convertUintTimestampToInt64(t *uint64) int64 {
-	if *t > uint64(math.MaxInt64) {
-		return int64(*t-math.MaxInt64) - int64(1)
-	}
-	return int64(*t) - math.MaxInt64 - int64(1)
 }
 
 func (self *Shard) getFieldsForSeries(db, series string, columns []string) ([]*metastore.Field, error) {