+++ integration/helpers/create_points.go
@@ -11,6 +11,18 @@ func CreatePoints(name string, numOfColumns, numOfPoints int) []*influxdb.Series
 	return CreatePointsFromFunc(name, numOfColumns, numOfPoints, func(int) float64 { return rand.Float64() })
 }
 
+func CreateSeries(prefix string, numOfSeries int) []*influxdb.Series {
+	s := make([]*influxdb.Series, numOfSeries)
+	for i := range s {
+		s[i] = &influxdb.Series{
+			Name:    fmt.Sprintf("%s%d", prefix, i),
+			Columns: []string{"column"},
+			Points:  [][]interface{}{{1}},
+		}
+	}
+	return s
+}
+
 func CreatePointsFromFunc(name string, numOfColumns, numOfPoints int, f func(int) float64) []*influxdb.Series {
 	series := &influxdb.Series{}