+++ CHANGELOG.md
@@ -1,10 +1,24 @@
 ## v0.9.0 [unreleased]
 
+### Features
+
+- [Issue #867](https://github.com/influxdb/influxdb/issues/867). Add option to return shard space mappings in list series
+
 ### Bugfixes
 
 - [Issue #652](https://github.com/influxdb/influxdb/issues/652). Return
   a meaningful error if an invalid column is used in where clause
   after joining multiple series
+
+## v0.8.2 [2014-09-08]
+
+### Features
+
+- Added API endpoint to update shard space definitions
+
+### Bugfixes
+
+- [Issue #886](https://github.com/influxdb/influxdb/issues/886). Shard space regexes reset after restart of InfluxDB
 
 ## v0.8.1 [2014-09-03]